Move type definitions to individual files.
