+++ src/lib/getIgnoredUpgrades.ts
@@ -1,6 +1,10 @@
 import { satisfies } from 'semver'
 import upgradePackageDefinitions from './upgradePackageDefinitions'
-import { IgnoredUpgrade, Index, Options, Version, VersionSpec } from '../types'
+import { IgnoredUpgrade } from '../types/IgnoredUpgrade'
+import { Index } from '../types/IndexType'
+import { Options } from '../types/Options'
+import { Version } from '../types/Version'
+import { VersionSpec } from '../types/VersionSpec'
 
 /** Get all upgrades that are ignored due to incompatible peer dependencies. */
 export async function getIgnoredUpgrades(
