+++ src/lib/getPeerDependencies.ts
@@ -2,7 +2,9 @@ import fs from 'fs'
 import jph from 'json-parse-helpfulerror'
 import { print } from '../logging'
 import getCurrentDependencies from './getCurrentDependencies'
-import { Index, Options, VersionSpec } from '../types'
+import { Index } from '../types/IndexType'
+import { Options } from '../types/Options'
+import { VersionSpec } from '../types/VersionSpec'
 
 /** Get peer dependencies from installed packages */
 function getPeerDependencies(current: Index<VersionSpec>, options: Options) {
