+++ src/lib/upgradePackageDefinitions.ts
@@ -4,7 +4,10 @@ import getPeerDependenciesFromRegistry from './getPeerDependenciesFromRegistry'
 import queryVersions from './queryVersions'
 import upgradeDependencies from './upgradeDependencies'
 import keyValueBy from './keyValueBy'
-import { Index, Options, VersionResult, VersionSpec } from '../types'
+import { Index } from '../types/IndexType'
+import { Options } from '../types/Options'
+import { VersionResult } from '../types/VersionResult'
+import { VersionSpec } from '../types/VersionSpec'
 
 /**
  * Returns an 3-tuple of upgradedDependencies, their latest versions and the resulting peer dependencies.
