+++ src/package-managers/index.ts
@@ -1,4 +1,5 @@
-import { Index, PackageManager } from '../types'
+import { Index } from '../types/IndexType'
+import { PackageManager } from '../types/PackageManager'
 import * as npm from './npm'
 import * as yarn from './yarn'
 import * as gitTags from './gitTags'
