+++ src/package-managers/yarn.ts
@@ -11,7 +11,13 @@ import spawn from 'spawn-please'
 import yaml from 'yaml'
 import * as versionUtil from '../version-util'
 import { viewOne, viewManyMemoized } from './npm'
-import { GetVersion, Index, Options, Packument, SpawnOptions, Version, NpmOptions } from '../types'
+import { GetVersion } from '../types/GetVersion'
+import { Index } from '../types/IndexType'
+import { Options } from '../types/Options'
+import { Packument } from '../types/Packument'
+import { SpawnOptions } from '../types/SpawnOptions'
+import { Version } from '../types/Version'
+import { NpmOptions } from '../types/NpmOptions'
 import { allowDeprecatedOrIsNotDeprecated, allowPreOrIsNotPre, satisfiesNodeEngine } from './filters'
 
 interface ParsedDep {
