+++ src/types/Maybe.ts
@@ -1,0 +1,2 @@
+/** A value that may be null or undefined. */
+export type Maybe<T = any> = T | null | undefined
