+++ src/types/NpmOptions.ts
@@ -1,0 +1,6 @@
+/** Options that can be provided to npm. */
+export interface NpmOptions {
+  global?: boolean
+  prefix?: string
+  registry?: string
+}
