+++ src/types/PackageFileRepository.ts
@@ -1,0 +1,5 @@
+/** Represents the repository field in package.json. */
+export interface PackageFileRepository {
+  url: string
+  directory?: string
+}
