+++ src/types/Packument.ts
@@ -1,0 +1,14 @@
+import { Index } from './IndexType'
+import { Version } from './Version'
+
+/** A pacote packument result object. */
+export interface Packument {
+  name: string
+  deprecated?: boolean
+  engines: {
+    node: string
+  }
+  time: Index<string>
+  version: Version
+  versions: Packument[]
+}
