+++ src/types/SpawnOptions.ts
@@ -1,0 +1,7 @@
+import { Index } from './IndexType'
+
+/** Options to the spawn node built-in. */
+export interface SpawnOptions {
+  env?: Index<string>
+  stderr?: (s: string) => void
+}
