+++ src/types/VersionLevel.ts
@@ -1,0 +1,2 @@
+/** The three main parts of a SemVer version. */
+export type VersionLevel = 'major' | 'minor' | 'patch'
