+++ src/types/VersionSpec.ts
@@ -1,0 +1,2 @@
+/** A version specification or range supported as a value in package.json dependencies. Includes SemVer, git urls, npm urls, etc. */
+export type VersionSpec = string
